+++ tests/feature/test_feature.py
@@ -73,7 +73,7 @@ class TestAttributes:
         dtypes = feature.dtypes
         assert isinstance(dtypes, dict)
         assert dtypes == {
-            "id": "int32",
+            "id": "int64",
             "x": "float64",
             "y": "float64",
         }
